Rename language file into Github repository.
